Refactor tasks into processors and validators
