+++ backend/contabilidad/utils/parsers/parser_libro_mayor.py
