+++ backend/contabilidad/utils/processors/base_processor.py
@@ -1,0 +1,5 @@
+import logging
+
+class BaseProcessor:
+    """Common utilities for processors."""
+    logger = logging.getLogger(__name__)
